Extract the data structure to store signed properties
